modified:   src/powadcr.cpp

+++ src/powadcr.cpp
@@ -6075,35 +6075,52 @@ void tapeControl() {
       //
       // Reproducimos el ultimo fichero cargado
       //
-      logln("State: Play was pressed. Loading last file.");
-      // Cargamos el ultimo fichero reproducido en el cassette
-      FILE_BROWSER_OPEN = false;
-      UPDATE_FROM_REMOTE_CONTROL = false;
-      ABORT = false;
-      EJECT = false;
-      PLAY = false;
-      REC = false;
-      STOP = true;  //26/04/2026 - Lo ponemos para que cuando cargue el ultimo fichero reproducido y después EJECT, 
-                    // se haga EJECT, porque la cinta está parada.
-      // Esto se usa para reproducir
-      FILE_SELECTED = true;
-      FILE_PREPARED = false;
-      LOADING_STATE = 0;
-      TAPESTATE = 100;
 
       // Cargamos el medio
       PATH_FILE_TO_LOAD = hmi.loadLastMedia();
-      // Cogemos el nombre del fichero
-      FILE_LOAD = getFileNameFromPath(PATH_FILE_TO_LOAD);
-      // Cogemos la ruta actual y la almacenamos
-      FILE_LAST_DIR = getDirFromPath(PATH_FILE_TO_LOAD);
-      // Cogemos el path anterior trabajando la cadena FILE_LAST_DIR
-      FILE_PREVIOUS_DIR = hmi.getPreviousDirFromPath(FILE_LAST_DIR); 
-      // Para el current dir que se muestra arriba. Ponemos un caracter " " al final como workaround, debido a un bug.
-      FILE_LAST_DIR_LAST = getDirFromPath(PATH_FILE_TO_LOAD) + " "; 
-      // Mostramos la página de carga
-      hmi.writeString("page tape");          
-      delay(125);
+
+      // Verificamos que PATH_FILE_TO_LOAD existe      
+      if (!SD_MMC.exists(PATH_FILE_TO_LOAD)) {
+        // En ese caso lanzamos un mensaje de error.
+        log_error("TAPE","File does not exist: " + PATH_FILE_TO_LOAD);
+        myNex.writeStr("tape0.g0.txt", "File already does not exist.");
+        PLAY = false;
+        STOP = true;
+        FILE_SELECTED = false;
+        FILE_PREPARED = false;
+        LOADING_STATE = 0;
+        TAPESTATE = 0;        
+      }
+      else
+      {
+        log_info("TAPE","State: Play was pressed. Loading last file.");
+        // Cargamos el ultimo fichero reproducido en el cassette
+        FILE_BROWSER_OPEN = false;
+        UPDATE_FROM_REMOTE_CONTROL = false;
+        ABORT = false;
+        EJECT = false;
+        PLAY = false;
+        REC = false;
+        STOP = true;  //26/04/2026 - Lo ponemos para que cuando cargue el ultimo fichero reproducido y después EJECT, 
+                      // se haga EJECT, porque la cinta está parada.
+        // Esto se usa para reproducir
+        FILE_SELECTED = true;
+        FILE_PREPARED = false;
+        LOADING_STATE = 0;
+        TAPESTATE = 100;
+
+        // Cogemos el nombre del fichero
+        FILE_LOAD = getFileNameFromPath(PATH_FILE_TO_LOAD);
+        // Cogemos la ruta actual y la almacenamos
+        FILE_LAST_DIR = getDirFromPath(PATH_FILE_TO_LOAD);
+        // Cogemos el path anterior trabajando la cadena FILE_LAST_DIR
+        FILE_PREVIOUS_DIR = hmi.getPreviousDirFromPath(FILE_LAST_DIR); 
+        // Para el current dir que se muestra arriba. Ponemos un caracter " " al final como workaround, debido a un bug.
+        FILE_LAST_DIR_LAST = getDirFromPath(PATH_FILE_TO_LOAD) + " ";
+        // Mostramos la página de carga
+        hmi.writeString("page tape");          
+        delay(125);
+      }
     } 
     else 
     {
